Modularize CloudSync database operations
